A Backup Script
This is the improved version of the fullBackup.sh. It uses the LZ4 algorithm to compress the "current" backup directory. Very low cpu usage and great filesize. It also removes the oldests backups from "backups" directory and send a 'say' command to the server (screen utility needed) when the compression starts and ends.